Add experimental JSON support for accept.
The JSON support here is meant as an example for future development. I
don't think the Jazz JSON API is stable enough to rely on at this point,
so all we're providing here is a basic HOWTO. In particular, the UUID
doesn't map to one of the (1001) style descriptive numbers anywhere and
the scm compare command doesn't provide the UUID for the changeset in
its JSON output. Neat, huh?